Reads rainfall data from a .txt file to generate IDF equation, and the information can be saved as a .pdf file.
The purpose of this plugin is to relate the rainfall station recorded in Brazil, with an IDF rainfall equation that describes it. From the location of this post, its historical series is combined with a probability distribution of best fit, resulting at the end of the process, in an Equation Intensity - Duration - Frequency of the respective post. This extension transforms this data into a file with the following information: I) Data of the Rainfall Post (name and identification code of the post, state and city where it is located, number of years, latitude and longitude), table with the rainfall data of the post that relates the year and its respective maximum precipitation in millimeters, II) Descriptive Statistics (mean, standard deviation, variance and the asymmetry coefficient), III) Probability Distribution of Adjustment and Isozone, IV) Information of the IDF Equation.